+++ chancy/contrib/__init__.py
@@ -1,0 +1,4 @@
+"""
+The contrib directory contains code that is not part of the core functionality
+of Chancy, but is useful for developers who are using Chancy.
+"""