+++ chancy/plugins/cron/django/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ChancyCronConfig(AppConfig):
+    name = "chancy.plugins.cron.django"
+    label = "chancy_cron"
+    verbose_name = "Chancy Cron"